TaskStorage class, which encapsulates a taskStorage object, is created in master branch.
