+++ src/main/java/Duke.java
@@ -3,30 +3,32 @@ import java.util.Collections;
 import java.util.ArrayList;
 public class Duke {
 
-    //private static Task[] taskstorage = new Task[101];
-    private static ArrayList<Task> taskstorage = new ArrayList<Task>(); // Create an ArrayList object
-    private static int ind = 0;
-    public static void addTask(Task t) {
-        System.out.println("Got it. I've added this task:");
-        System.out.println("  " + t);
-        taskstorage.add(t);
-        ind++;
-        System.out.println("Now you have " + ind + " task(s) in the list.");
-    }
-
-    public static void deleteTask(Task t) {
-        System.out.println("Noted. I've removed this task:");
-        System.out.println("  " + t);
-        taskstorage.remove(t);
-        ind--;
-        System.out.println("Now you have " + ind + " task(s) in the list.");
-    }
+//    //private static Task[] taskstorage = new Task[101];
+//    private static ArrayList<Task> taskstorage = new ArrayList<Task>();
+//    // Tasks are indexed from 0 in taskstorage.
+//    private static int ind = 0; //Number of Current Tasks
+//    public static void addTask(Task t) {
+//        System.out.println("Got it. I've added this task:");
+//        System.out.println("  " + t);
+//        taskstorage.add(t);
+//        ind++;
+//        System.out.println("Now you have " + ind + " task(s) in the list.");
+//    }
+//
+//    public static void deleteTask(Task t) {
+//        System.out.println("Noted. I've removed this task:");
+//        System.out.println("  " + t);
+//        taskstorage.remove(t);
+//        ind--;
+//        System.out.println("Now you have " + ind + " task(s) in the list.");
+//    }
 
     public static void main(String[] args) throws IOException, DukeException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         PrintWriter pw = new PrintWriter(new BufferedWriter(new OutputStreamWriter(System.out)));
         StringBuilder sb = new StringBuilder();
-
+        //TaskManagement taskManager = new TaskManagement(); // to manage saved data
+        TaskStorage taskstorage = new TaskStorage();
         String logo = " ____        _        \n"
                 + "|  _ \\ _   _| | _____ \n"
                 + "| | | | | | | |/ / _ \\\n"
@@ -43,7 +45,6 @@ public class Duke {
          * Simply echoes commands entered by the user,
          * and exits when the user types "bye".
          */
-        //int ind = 1;
         while (true) {
             inp = br.readLine();
             String[] input = inp.split(" ");
@@ -52,9 +53,9 @@ public class Duke {
                 case "list":
                     System.out.println("Tasks:");
                     //System.out.println(taskstorage);
-                    for (int i = 0; i < ind; i++) {
+                    for (int i = 0; i < taskstorage.noTasks(); i++) {
                         System.out.print(i + 1 + ".");
-                        System.out.println(taskstorage.get(i));
+                        System.out.println(taskstorage.getTask(i));
                     }
                     break;
 
@@ -66,10 +67,10 @@ public class Duke {
                     //System.out.println("Nice! I've marked this task as done:");
                     try {
                         int taskNo = Integer.parseInt(input[1]);
-                        if (taskNo >= ind || taskNo <= 0) {
+                        if (taskNo > taskstorage.noTasks() || taskNo <= 0) {
                             throw new DukeException("Give a vaild number");
                         }
-                        taskstorage.get(taskNo - 1).markasDone();
+                        taskstorage.getTask(taskNo - 1).markasDone();
                     } catch (NumberFormatException e) {
                         System.out.println("Number should be typed in");
                     } catch (DukeException e){
@@ -81,10 +82,10 @@ public class Duke {
                 case "unmark":
                     try {
                         int taskNoUnmark = Integer.parseInt(input[1]);
-                        if (taskNoUnmark >= ind || taskNoUnmark <= 0) {
+                        if (taskNoUnmark > taskstorage.noTasks() || taskNoUnmark <= 0) {
                             throw new DukeException("Give a valid number");
                         }
-                        taskstorage.get(taskNoUnmark - 1).markasUnDone();
+                        taskstorage.getTask(taskNoUnmark - 1).markasUnDone();
                     } catch (NumberFormatException e) {
                         System.out.println("Number should be typed in");
                     } catch (DukeException e) {
@@ -101,7 +102,7 @@ public class Duke {
                         }
                         String todoTask = inp.substring(5);
                         Task todo = new Todo(todoTask);
-                        addTask(todo);
+                        taskstorage.addTask(todo);
                         break;
                     } catch (DukeException e) {
                         System.out.println(e.getMessage());
@@ -123,7 +124,7 @@ public class Duke {
                         String deadLineTaskStr = inputDeadline[0];
                         String end = inputDeadline[1].substring(3);
                         Task deadLineTask = new Deadline(deadLineTaskStr, end);
-                        addTask(deadLineTask);
+                        taskstorage.addTask(deadLineTask);
                     } catch (DukeException e) {
                         System.out.println(e.getMessage());
                     } finally {
@@ -146,7 +147,7 @@ public class Duke {
                         String eventBegin = eventStrsplit[1].substring(5);
                         String eventEnd = eventStrsplit[2].substring(3);
                         Task eventTask = new Event(eventTaskStr, eventBegin, eventEnd);
-                        addTask(eventTask);
+                        taskstorage.addTask(eventTask);
                     } catch (DukeException e) {
                         System.out.println(e.getMessage());
                     } finally {
@@ -156,11 +157,11 @@ public class Duke {
                 case "delete":
                     try {
                         int taskNo = Integer.parseInt(input[1]);
-                        if (taskNo > ind || taskNo <= 0) {
+                        if (taskNo > taskstorage.noTasks() || taskNo <= 0) {
                             throw new DukeException("Give a vaild number");
                         }
-                        Task eventTask = taskstorage.get(taskNo - 1);
-                        deleteTask(eventTask);
+                        Task eventTask = taskstorage.getTask(taskNo - 1);
+                        taskstorage.deleteTask(eventTask);
                     } catch (NumberFormatException e) {
                         System.out.println("Number should be typed in");
                     } catch (DukeException e){
@@ -172,7 +173,10 @@ public class Duke {
                 default:
                     DukeException dukeException = new DukeException();
                     System.out.println(dukeException.getMessage());
+
+
             }
+            //taskManager.save(taskstorage);
             System.out.println(line);
             if (inp.equals("bye")) {
                 break;
